refactor: move the request logic to service folder
